refatoracao: mover validacao de exclusao para service

+++ src/services/usuario.service.js
@@ -1,5 +1,6 @@
 const bcrypt = require("bcrypt")
 const Usuario = require("../model/usuario.model")
+const AppError = require("../utils/AppError")
 
 async function criar(nome,email,senha){
     const usuarioExiste = await Usuario.findOne({ email })
@@ -23,7 +24,12 @@ async function listar(){
 }
 
 async function deletar(id){
-    const usuario = await Usuario.findByIdAndDelete(id)
+    const usuario = await Usuario.findById(id)
+
+  if(!usuario){
+           throw new AppError("usuário não encontrado", 404)
+           }
+      await Usuario.findByIdAndDelete(id)
 
     return usuario
 }
@@ -59,5 +65,6 @@ module.exports = {
     criar,
     deletar,
     perfil,
-    atualizar
+    atualizar,
+    listar
 }
